Fix the missing SplitView divider thickness.

+++ Source/Custom Views/MISplitView.h
@@ -19,15 +19,15 @@
 #import <Foundation/Foundation.h>
 
 @interface MISplitView : NSSplitView {
-  float thickness;
+  CGFloat thickness;
   BOOL drawLowerBorder;
 }
 
 - (void)saveLayoutUsingName:(NSString*)name;
 - (void)restoreLayoutUsingName:(NSString*)name;
 
-- (void)setDividerThickness:(float)thickness;
-- (float)dividerThickness;
+- (void)setDividerThickness:(CGFloat)thickness;
+- (CGFloat)dividerThickness;
 
 - (void)setDrawLowerBorder:(BOOL)flag;
 - (BOOL)drawLowerBorder;
